rearrange the App.tsx and move rootStack to a new file and apply useContext in the App.tsx
