+++ package.json
@@ -50,6 +50,7 @@
     "@types/express": "^4.17.17",
     "@types/node": "^18.14.6",
     "@types/react": "~18.0.14",
+    "@types/react-native": "^0.71.3",
     "@types/react-native-dotenv": "^0.2.0",
     "nodemon": "^2.0.21",
     "react-native-dotenv": "^3.4.8",
