+++ src/screens/Community.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import MainScreen from '../components/community/Main';
@@ -5,7 +7,7 @@ import AddPostScreen from '../components/community/AddPost';
 import CommentScreen from '../components/community/Comment';
 
 export type CommunityStackParamList = {
-  Community: undefined;
+  Main: undefined;
   AddPost: undefined;
   Comment: undefined;
 };
@@ -16,7 +18,7 @@ export default function CommunityScreen() {
 
   return (
     <Stack.Navigator screenOptions={{ headerShown: false}}>
-      <Stack.Screen name="Community" component={MainScreen} />
+      <Stack.Screen name="Main" component={MainScreen} />
       <Stack.Screen name="AddPost" component={AddPostScreen} />
       <Stack.Screen name="Comment" component={CommentScreen} />
     </Stack.Navigator>
