+++ src/screens/Home.tsx
@@ -1,11 +1,14 @@
 import * as React from 'react';
+import { useContext } from 'react';
+import { UserContext } from '../../App';
 import { View, Text, Image, Button, TouchableOpacity, ScrollView} from 'react-native';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { useNavigation } from '@react-navigation/native';
 
 import { StyleSheet } from 'react-native';
 import Styles from '../constants/Styles';
-import { RootStackParamList } from '../../App';
+import { RootStackParamList } from '../../RootStack';
+
 
 type HomeScreenNavigationProp = NativeStackNavigationProp<
   RootStackParamList,
@@ -47,6 +50,8 @@ const tempStyling = StyleSheet.create({
 })
 export default function HomeScreen() {
   const navigation = useNavigation<HomeScreenNavigationProp>();
+  const { user } = useContext(UserContext);
+  console.log('user info in home: ', user);
   return (
     <>
       <ScrollView style={{backgroundColor: 'white', flexDirection: 'column'}}>
