+++ src/screens/ProfilePlaceholder.tsx
@@ -4,7 +4,7 @@ import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { useNavigation } from '@react-navigation/native';
 
 import Styles from '../constants/Styles';
-import { RootStackParamList } from '../../App';
+import { RootStackParamList } from '../../RootStack';
 
 type ProfilePlaceholderNavigationProp = NativeStackNavigationProp<RootStackParamList>;
