+++ src/screens/SignUp.tsx
@@ -1,32 +1,21 @@
-import React, { useState, useEffect, useContext } from 'react';
-import {
-  View,
-  SafeAreaView,
-  Text,
-  Button,
-  TouchableOpacity,
-  CheckBox,
-  TextInput,
-  StyleSheet,
-} from 'react-native';
-import {
-  NativeStackNavigationProp,
-  NativeStackScreenProps,
-} from '@react-navigation/native-stack';
+import React, { useState, useEffect, useContext, CheckBox } from 'react';
+import { View, SafeAreaView, Text, Button, TouchableOpacity, TextInput, StyleSheet} from 'react-native';
+import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { useNavigation } from '@react-navigation/native';
 import Styles from '../constants/Styles';
-import { RootStackParamList } from '../../App';
-// import FormTextInput from "../components/Signup/TextInput";
+import { RootStackParamList } from '../../RootStack';
 import { auth } from '../constants/firebase/firebase';
 import { createUserWithEmailAndPassword } from 'firebase/auth';
 import axios from 'axios';
 import { UserContext } from '../../App';
 
-type MessagesScreenNavigationProp =
+type SignUpScreenNavigationProp =
   NativeStackNavigationProp<RootStackParamList>;
 
 const SignUpScreen = () => {
-  const navigation = useNavigation<MessagesScreenNavigationProp>();
+  const navigation = useNavigation<SignUpScreenNavigationProp>();
+
+  const { user, setUser } = useContext(UserContext);
 
   const [userInformation, setUserInformation] = useState({
     firstName: '',
@@ -37,15 +26,14 @@ const SignUpScreen = () => {
     householdID: null,
   });
   const [selectNotifications, setSelectNotifications] = useState(false);
-  const { setUserID, userID } = useContext(UserContext); //set USERID after axios call
 
   useEffect(() => {
     console.log(userInformation, selectNotifications);
   }, [userInformation, selectNotifications]);
 
   useEffect(() => {
-    console.log('userID', userID);
-  }, [userID]);
+    console.log('userID', user?.id);
+  }, [user]);
 
   const createUser = () => {
     createUserWithEmailAndPassword(
@@ -70,8 +58,8 @@ const SignUpScreen = () => {
           .catch((error) => {
             console.log(error);
           });
-        setUserID({
-          userId: userID,
+        setUser({
+          id: userID,
           firstName: userInformation.firstName,
           lastName: userInformation.lastName,
           email: userInformation.email,
@@ -88,109 +76,106 @@ const SignUpScreen = () => {
   };
 
   return (
-    <>
-      <SafeAreaView style={Styles.container}>
-        <Button title="Back" onPress={() => navigation.navigate('Home')} />
-        <View>
-          <Text style={styles.headertext}>Sign Up</Text>
-        </View>
-        {/* <FormTextInput
+    <View style={Styles.container}>
+      <Button title="Back" onPress={() => navigation.navigate('Home')} />
+      <View>
+        <Text style={styles.headertext}>Sign Up</Text>
+      </View>
+      {/* <FormTextInput
           label="First Name"
           error="Please enter a Name"
           password={false}
           placeholder="Enter your Name"
           }}
         ></FormTextInput> */}
-        <View>
-          <TouchableOpacity>
-            <Text>First Name</Text>
-            <TextInput
-              placeholder="Enter your First Name"
-              onChangeText={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['firstName']: text,
-                });
-              }}
-              required
-              style={styles.formInput}
-            ></TextInput>
-            <Text>Last Name</Text>
-            <TextInput
-              placeholder="Enter your Last Name"
-              onChangeText={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['lastName']: text,
-                });
-              }}
-              style={styles.formInput}
-            ></TextInput>
-            <Text>Email</Text>
-            <TextInput
-              placeholder="Enter your Email"
-              type="email"
-              required
-              onChangeText={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['email']: text,
-                });
-              }}
-              style={styles.formInput}
-            ></TextInput>
-            <Text>Password</Text>
-            <TextInput
-              placeholder="Enter your Password"
-              secureTextEntry={true}
-              onChangeText={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['password']: text,
-                });
-              }}
-              style={styles.formInput}
-            ></TextInput>
-            <Text>Re-enter Password</Text>
-            <TextInput
-              placeholder="Confirm Password"
-              secureTextEntry={true}
-              onChangeText={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['confirmPassword']: text,
-                });
-              }}
-              style={styles.formInput}
-            ></TextInput>
-            <Text>Enable Notifications?</Text>
-            <CheckBox
+      <View>
+        <TouchableOpacity>
+          <Text>First Name</Text>
+          <TextInput
+            placeholder="Enter your First Name"
+            onChangeText={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['firstName']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Text>Last Name</Text>
+          <TextInput
+            placeholder="Enter your Last Name"
+            onChangeText={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['lastName']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Text>Email</Text>
+          <TextInput
+            placeholder="Enter your Email"
+            type="email"
+            required
+            onChangeText={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['email']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Text>Password</Text>
+          <TextInput
+            placeholder="Enter your Password"
+            secureTextEntry={true}
+            onChangeText={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['password']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Text>Re-enter Password</Text>
+          <TextInput
+            placeholder="Confirm Password"
+            secureTextEntry={true}
+            onChangeText={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['confirmPassword']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Text>Enable Notifications?</Text>
+          {/* <CheckBox
               value={selectNotifications}
               onValueChange={setSelectNotifications}
               style={styles.checkbox}
-            ></CheckBox>
-            <Text>Household ID (Optional)</Text>
-            <TextInput
-              placeholder="Household ID (Optional)"
-              onChangeNumber={(text) => {
-                setUserInformation({
-                  ...userInformation,
-                  ['householdID']: text,
-                });
-              }}
-              style={styles.formInput}
-            ></TextInput>
-            <Button onPress={createUser} title="Submit"></Button>
-          </TouchableOpacity>
-        </View>
-      </SafeAreaView>
-    </>
+            ></CheckBox> */}
+          <Text>Household ID (Optional)</Text>
+          <TextInput
+            placeholder="Household ID (Optional)"
+            onChangeNumber={(text) => {
+              setUserInformation({
+                ...userInformation,
+                ['householdID']: text,
+              });
+            }}
+            style={styles.formInput}
+          ></TextInput>
+          <Button onPress={createUser} title="Submit"></Button>
+        </TouchableOpacity>
+      </View>
+    </View>
   );
 };
 
 const styles = {
   headertext: {
-    fontSize: '15vw',
+    fontSize: 15,
   },
   checkbox: {
     alignSelf: 'flex-start',
@@ -198,8 +183,8 @@ const styles = {
   formInput: {
     border: 'solid',
     textcolor: 'lightgrey',
-    padding: '10px',
-    buffer: '5px',
+    padding: 10,
+    buffer: 5,
     // padding-bottom: '10px',
   },
 };
